Test updates

+++ test/integration/authz/cli/cli_helpers.bash
@@ -60,18 +60,11 @@ checkInvariant() {
  return 0
 }
 notAuthorized() {
-	NOTAUTHORIZED="Error response from daemon: Not Authorized!"
-#	return 0  #temporary fix to all tests to run until we support all the container commands
-#	run docker -H $SWARM_HOST --config $1 attach $2	
-#   if !( [ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]] ); then
-#       return 200
-#   fi
-#   if !([ "$status" -ne 0 ] ); then
-#        return 1
-#    fi
-#    if !([[ "$output" == *"$NOTAUTHORIZED"* ]]); then
-#        return 100
-#    fi
+   NOTAUTHORIZED="Error response from daemon: Not Authorized!"
+   run docker -H $SWARM_HOST --config $1 attach $2	
+   if !( [ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]] ); then
+       return 1
+   fi
 #	run docker -H $SWARM_HOST --config $1 exec $2 ls
 #    if !( [ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]] ); then
 #        return 2
@@ -100,10 +93,10 @@ notAuthorized() {
 #    if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
 #        return 10
 #    fi
-#	run docker -H $SWARM_HOST --config $1 restart $2
-#   if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
-#       return 11
-#   fi
+	run docker -H $SWARM_HOST --config $1 restart $2
+    if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
+       return 11
+    fi
 	run docker -H $SWARM_HOST --config $1 rm $2
     if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
         return 12
@@ -112,14 +105,14 @@ notAuthorized() {
     if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
         return 13
     fi
-#	run docker -H $SWARM_HOST --config $1 stats $2
-#   if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
-#        return 14
-#   fi
-#    run docker -H $SWARM_HOST --config $1 top $2
-#    if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
-#        return 15
-#    fi
+	run docker -H $SWARM_HOST --config $1 stats $2
+    if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
+        return 14
+    fi
+    run docker -H $SWARM_HOST --config $1 top $2
+    if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
+        return 15
+    fi
 	run docker -H $SWARM_HOST --config $1 unpause $2
     if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
         return 16
@@ -128,10 +121,10 @@ notAuthorized() {
     if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
         return 17
     fi
-#    run docker -H $SWARM_HOST --config $1 wait $2
-#    if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
-#       return 18
-#    fi
+    run docker -H $SWARM_HOST --config $1 wait $2
+    if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
+       return 18
+    fi
 	run docker -H $SWARM_HOST --config $1 stop $2
     if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
         return 19
@@ -140,11 +133,6 @@ notAuthorized() {
     if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
         return 20
     fi
-	run docker -H $SWARM_HOST --config $1 pause $2
-    if !([ "$status" -ne 0 ] && [[ "$output" == *"$NOTAUTHORIZED"* ]]); then
-        return 21
-    fi
-	#'
 	return 0
 }
